Test User Service

+++ src/main/java/com/example/productcategoryservice/service/UserService.java
@@ -20,21 +20,23 @@ public class UserService {
     }
 
     public void save(User user) {
+        if (user == null) {
+            throw new RuntimeException("User can't be null");
+        }
         userRepository.save(user);
     }
 
     public void deleteById(int id) {
+        if (userRepository.findById(id).isEmpty()){
+            throw new RuntimeException("User can not a found");
+        }
         userRepository.deleteById(id);
 
     }
 
     public void register(User user) {
         userRepository.save(user);
-
     }
 
-    public void edit(UserSaveDto sa){
-
-    }
 
 }
